feat: add FreeBSD platform support to the build system and source tree
Add CMake platform detection and conditionals for FreeBSD, kqueue-based
I/O backend, platform backends for memory/sysinfo/system, and portability
fixes for clang/libc++ (chrono timezone shims, atomic_shared_ptr,
move_only_function, enumerate utility).

Adapt C++20 module builds so that jsonl and parquet readers compile as
regular translation units on FreeBSD via header fallbacks, and let cc-rs
fall back to the system compiler when no explicit toolchain root is set.

+++ src/query/interpret/awesome_memgraph_functions.cpp
@@ -1488,7 +1488,7 @@ TypedValue Timestamp(const TypedValue *args, int64_t nargs, const FunctionContex
     return TypedValue(arg.ValueDuration().microseconds, ctx.memory);
   }
   if (arg.IsZonedDateTime()) {
-    return TypedValue(arg.ValueZonedDateTime().SysMicrosecondsSinceEpoch().count(), ctx.memory);
+    return TypedValue(static_cast<int64_t>(arg.ValueZonedDateTime().SysMicrosecondsSinceEpoch().count()), ctx.memory);
   }
   return TypedValue(ctx.timestamp, ctx.memory);
 }
